demrepofdave/al5int/add_roms_and_code_improvements: Fixed allegro4 build. Moved keybpoard routines back to electron logic
