+++ Makefile.am
@@ -70,6 +70,7 @@ elkulator_SOURCES = src/1770.c \
                     src/disc.c \
                     src/fdi2raw.c \
                     src/fdi.c \
+                    src/keyboard.c \
                     src/logger.c \
                     src/main.c \
                     src/mem.c \
