+++ src/callback_handlers.h
@@ -3,8 +3,12 @@
 
 #include <stdint.h>
 
+#include "keyboard.h"
+
 typedef void (*callback_filename_t)(const char * filename);
+typedef void (*callback_keyboard_t)(host_key_t hostkey);
 typedef void (*callback_void_t)(void);
+
 
 
 typedef struct 
@@ -32,6 +36,10 @@ typedef struct
     callback_filename_t handle_startmovie;
     callback_void_t     handle_stopmovie;
     callback_void_t     handle_enable_debugger;
+
+    // Keyboard callback handlers
+    callback_keyboard_t handle_key_down;
+    callback_keyboard_t handle_key_up;
 
 } callback_handlers_t;
