+++ src/host_abstraction_layer/allegro_4/video.c
@@ -292,3 +292,18 @@ void video_shutdown()
 {
     allegro_exit();
 }
+
+void video_start_timer()
+{
+    // For allegro4 there is nothing to do.
+}
+
+void video_stop_timer()
+{
+    // For allegro4 there is nothing to do.
+}
+
+void video_set_window_title(const char * title)
+{
+    // TODO: Allegro4 version.
+}
