--- src/host_abstraction_layer/allegro_5/linux-keydefine.c
@@ -1,70 +1,0 @@
-/*Elkulator v1.0 by Sarah Walker
-  Linux keyboard redefinition GUI*/
-
-#include <allegro5/allegro.h>
-#include "elk.h"
-#include "host_abstraction_layer/keyboard.h"
-#include "keyboard_internal.h"
-#include "config_vars.h"
-
-extern bool elk_key_state[ELK_KEY_MAX];
-
-char *key_names[] =
-{
-   "",           "A",          "B",          "C",
-   "D",          "E",          "F",          "G",
-   "H",          "I",          "J",          "K",
-   "L",          "M",          "N",          "O",
-   "P",          "Q",          "R",          "S",
-   "T",          "U",          "V",          "W",
-   "X",          "Y",          "Z",          "0",
-   "1",          "2",          "3",          "4",
-   "5",          "6",          "7",          "8",
-   "9",          "0_PAD",      "1_PAD",      "2_PAD",
-   "3_PAD",      "4_PAD",      "5_PAD",      "6_PAD",
-   "7_PAD",      "8_PAD",      "9_PAD",      "F1",
-   "F2",         "F3",         "F4",         "F5",
-   "F6",         "F7",         "F8",         "F9",
-   "F10",        "F11",        "F12",        "ESC",
-   "TILDE",      "MINUS",      "EQUALS",     "BACKSPACE",
-   "TAB",        "OPENBRACE",  "CLOSEBRACE", "ENTER",
-   "COLON",      "QUOTE",      "BACKSLASH",  "BACKSLASH2",
-   "COMMA",      "STOP",       "SLASH",      "SPACE",
-   "INSERT",     "DEL",        "HOME",       "END",
-   "PGUP",       "PGDN",       "LEFT",       "RIGHT",
-   "UP",         "DOWN",       "SLASH_PAD",  "ASTERISK",
-   "MINUS_PAD",  "PLUS_PAD",   "DEL_PAD",    "ENTER_PAD",
-   "PRTSCR",     "PAUSE",      "ABNT_C1",    "YEN",
-   "KANA",       "CONVERT",    "NOCONVERT",  "AT",
-   "CIRCUMFLEX", "COLON2",     "KANJI",      "EQUALS_PAD",
-   "BACKQUOTE",  "SEMICOLON",  "COMMAND",    "UNKNOWN1",
-   "UNKNOWN2",   "UNKNOWN3",   "UNKNOWN4",   "UNKNOWN5",
-   "UNKNOWN6",   "UNKNOWN7",   "UNKNOWN8",   "LSHIFT",
-   "RSHIFT",     "LCONTROL",   "RCONTROL",   "ALT",
-   "ALTGR",      "LWIN",       "RWIN",       "MENU",
-   "SCRLOCK",    "NUMLOCK",    "CAPSLOCK",   "MAX"
-};
-
-/* Key reading control. */
-static bool special_key_pressed(elk_key_id_t elk_keycode)
-{
-    bool result = false;
-    for (int i = 0; i < HOST_KEY_MAX; i++)
-    {
-        if (elkConfig.keyboard.host_key_mapping[i] == elk_keycode && elk_key_state[i])
-        {
-            result = true;
-        }
-    }
-    return result;
-}
-
-bool break_pressed()
-{
-    return special_key_pressed(ELK_KEY_BREAK);
-}
-
-bool menu_pressed()
-{
-    return special_key_pressed(ELK_SPECIAL_KEY_MENU);
-}
