+++ src/host_abstraction_layer/common/keyutils.c
@@ -1,211 +1,7 @@
 
 #include <string.h>
-#include "host_abstraction_layer/keyboard.h"
+#include "keyboard.h"
 #include "logger.h"
-
-typedef struct 
-{
-    const char * config_key_string;
-    const char * key_longname;
-} key_strings_t;
-
-static key_strings_t elk_keycode_config_string[ELK_KEY_MAX] =
-{
-    { "",                     ""  },
-    { "elk_key_0",            "0" },
-    { "elk_key_1",            "1" },
-    { "elk_key_2",            "2" },
-    { "elk_key_3",            "3" },
-    { "elk_key_4",            "4" },
-    { "elk_key_5",            "5" },
-    { "elk_key_6",            "6" },
-    { "elk_key_7",            "7" },
-    { "elk_key_8",            "8" },
-    { "elk_key_9",            "9" },
-    { "elk_key_a",            "A" },
-    { "elk_key_b",            "B" },
-    { "elk_key_c",            "C" },
-    { "elk_key_d",            "D" },
-    { "elk_key_e",            "E" },
-    { "elk_key_f",            "F" },
-    { "elk_key_g",            "G" },
-    { "elk_key_h",            "H" },
-    { "elk_key_i",            "I" },
-    { "elk_key_j",            "J" },
-    { "elk_key_k",            "K" },
-    { "elk_key_l",            "L" },
-    { "elk_key_m",            "M" },
-    { "elk_key_n",            "N" },
-    { "elk_key_o",            "O" },
-    { "elk_key_p",            "P" },
-    { "elk_key_q",            "Q" },
-    { "elk_key_r",            "R" },
-    { "elk_key_s",            "S" },
-    { "elk_key_t",            "T" },
-    { "elk_key_u",            "U" },
-    { "elk_key_v",            "V" },
-    { "elk_key_w",            "W" },
-    { "elk_key_x",            "X" },
-    { "elk_key_y",            "Y" },
-    { "elk_key_z",            "Z" },
-    { "elk_key_equals",       "=" },
-    { "elk_key_comma",        "," },
-    { "elk_key_fullstop",     "." },
-    { "elk_key_slash",        "/" },
-    { "elk_key_semicolon",    ";" },
-    { "elk_key_colon",        ":" },
-    { "elk_key_left",         "Left" },
-    { "elk_key_right",        "Right" },
-    { "elk_key_up",           "Up" },
-    { "elk_key_down",         "Down" },
-    { "elk_key_function",     "Function" },
-    { "elk_key_copy",         "Copy" },
-    { "elk_key_control",      "Control" },
-    { "elk_key_shift",        "Shift" },
-    { "elk_key_del",          "Delete" },
-    { "elk_key_space",        "Spacebar" },
-    { "elk_key_return",       "Return" },
-    { "elk_key_escape",       "Escape" },
-    { "elk_key_break",        "Break" },
-    { "elk_special_key_menu", "<menu>" }
-};
-
-
-// see https://planet.racket-lang.org/package-source/kazzmir/allegro.plt/1/6/allegro-4.2.0/examples/exkeys.c
-static const key_strings_t host_key_string_table[HOST_KEY_MAX] = 
-{
-    { "", "" },
-    { "host_key_a", "A" },
-    { "host_key_b", "B" },
-    { "host_key_c", "C" },
-    { "host_key_d", "D" },
-    { "host_key_e", "E" },
-    { "host_key_f", "F" },
-    { "host_key_g", "G" },
-    { "host_key_h", "H" },
-    { "host_key_i", "I" },
-    { "host_key_j", "J" },
-    { "host_key_k", "K" },
-    { "host_key_l", "L" },
-    { "host_key_m", "M" },
-    { "host_key_n", "N" },
-    { "host_key_o", "O" },
-    { "host_key_p", "P" },
-    { "host_key_q", "Q" },
-    { "host_key_r", "R" },
-    { "host_key_s", "S" },
-    { "host_key_t", "T" },
-    { "host_key_u", "U" },
-    { "host_key_v", "V" },
-    { "host_key_w", "W" },
-    { "host_key_x", "X" },
-    { "host_key_y", "Y" },
-    { "host_key_z", "Z" },
-
-    { "host_key_0", "0" },
-    { "host_key_1", "1" },
-    { "host_key_2", "2" },
-    { "host_key_3", "3" },
-    { "host_key_4", "4" },
-    { "host_key_5", "5" },
-    { "host_key_6", "6" },
-    { "host_key_7", "7" },
-    { "host_key_8", "8" },
-    { "host_key_9", "9" },
-
-    { "host_key_pad_0",        "Keypad 0" },
-    { "host_key_pad_1",        "Keypad 1" },
-    { "host_key_pad_2",        "Keypad 2" },
-    { "host_key_pad_3",        "Keypad 3" },
-    { "host_key_pad_4",        "Keypad 4" },
-    { "host_key_pad_5",        "Keypad 5" },
-    { "host_key_pad_6",        "Keypad 6"  },
-    { "host_key_pad_7",        "Keypad 7" },
-    { "host_key_pad_8",        "Keypad 8" },
-    { "host_key_pad_9",        "Keypad 9"  },
-
-    { "host_key_f1",  "F1"   },
-    { "host_key_f2",  "F2"   },
-    { "host_key_f3",  "F3"   },
-    { "host_key_f4",  "F4"   },
-    { "host_key_f5",  "F5"   },
-    { "host_key_f6",  "F6"   },
-    { "host_key_f7",  "F7"   },
-    { "host_key_f8",  "F8"   },
-    { "host_key_f9",  "F9"   },
-    { "host_key_f10", "F10"   },
-    { "host_key_f11", "F11"   },
-    { "host_key_f12", "F12"   },
-
-    { "host_key_escape",      "Escape"        },
-    { "host_key_tilde",       "` (tilde)"     },
-    { "host_key_minus",       "-"             },
-    { "host_key_equals",      "="             },
-    { "host_key_backspace",   "Backspace"     },
-
-    { "host_key_tab",         "Tab"           },
-    { "host_key_openbrace",   "["             },
-    { "host_key_closebrace",  "]"             },
-    { "host_key_enter",       "Enter"         },
-    { "host_key_semicolon",   ";"             },
-    { "host_key_apostrophe",  "'"             },
-    { "host_key_backslash",   "\\"            },
-    { "host_key_backslash2",  "\\2"           },
-    { "host_key_comma",       ","             },
-    { "host_key_fullstop",    "."             },
-    { "host_key_slash",       "/"             },
-    { "host_key_space",       "Space"         },
-    { "host_key_insert",      "Insert"        },
-    { "host_key_delete",      "Delete"        },
-    { "host_key_home",        "Home"          },
-    { "host_key_end",         "End"           },
-    { "host_key_pageup",      "Page Up"       },
-    { "host_key_pagedown",    "Page Down"     },
-
-    { "host_key_left",        "Arrow Left"    },
-    { "host_key_right",       "Arrow Right"   },
-    { "host_key_up",          "Arrow Up"      },
-    { "host_key_down",        "Arror Down"    },
-
-    { "host_key_pad_slash",    "Keypad /"  },
-    { "host_key_pad_asterisk", "Keypad *"  },
-    { "host_key_pad_minus",    "Keypad -"  },
-    { "host_key_pad_plus",     "Keypad +"  },
-    { "host_key_pad_delete",   "Keypad Del"   },
-    { "host_key_pad_enter",    "Keypad Enter"  },
-    { "host_key_print_screen", "Print Screen"  },
-    { "host_key_pause",        "Pause"  },
-
-    { "host_key_abnt_c1",      "AbntC1"  },
-    { "host_key_yen",          "Yen"     },
-    { "host_key_kana",         "Kana"    },
-    { "host_key_convert",      "Convert"    },
-    { "host_key_noconvert",    "Noconvert"    },
-
-    { "host_key_at",           "AT"    },
-    { "host_key_circumflex",   "Circumflex" },
-    { "host_key_colon2",       "Colon2"    },
-    { "host_key_kanji",        "Kanji"    },
-    { "host_key_pad_equals",   "Keypad ="    },
-    { "host_key_backquote",    "Backquote"    },
-    { "host_key_semicolon2",   "Semicolon2"    },
-    { "host_key_command",      "Command"    },
-
-    { "host_key_lshift",      "Left Shift"    },
-    { "host_key_rshift",      "Right Shift"   },
-    { "host_key_lctrl",       "Left CTRL"     },
-    { "host_key_rctrl",       "Left CTRL"     },
-    { "host_key_alt",         "ALT"           },
-    { "host_key_altgr",       "ALTGR"         },
-
-    { "host_key_lwin",        "Left WIN"      },
-    { "host_key_rwin",        "Right WIN"     },
-    { "host_key_menu",        "MENU"          },
-    { "host_key_scroll_lock", "Scroll Lock"   },
-    { "host_key_num_lock",    "Num Lock"      },
-    { "host_key_caps_lock",   "Caps Lock"     },
-};
-
 
 // see https://planet.racket-lang.org/package-source/kazzmir/allegro.plt/1/6/allegro-4.2.0/examples/exkeys.c
 static const int old_config_keys_to_new_host_keys_table[128] =
@@ -367,63 +163,4 @@ host_key_t keyutils_get_hostkey_from_legacy_keyid(uint8_t old_key_id)
     return(host_key);
 }
 
-const char * keyutils_get_hostkey_longname(host_key_t host_key)
-{
-    const char * config_str = NULL;
-    if(host_key < HOST_KEY_MAX)
-    {
-        config_str = host_key_string_table[host_key].key_longname;
-    }
-    return config_str;
-}
 
-
-const char * keyutils_get_hostkey_config_string(host_key_t host_key)
-{
-    const char * config_str = NULL;
-    if(host_key < HOST_KEY_MAX)
-    {
-        config_str = host_key_string_table[host_key].config_key_string;
-    }
-    return config_str;
-}
-
-const char * keyutils_get_elkkey_longname(elk_key_id_t elk_key)
-{
-    const char * config_str = NULL;
-    if(elk_key < ELK_KEY_MAX)
-    {
-        config_str = elk_keycode_config_string[elk_key].key_longname;
-    }
-    return config_str;
-}
-
-const char * keyutils_get_elkkey_config_string(elk_key_id_t elk_key)
-{
-    const char * config_str = NULL;
-    if(elk_key < ELK_KEY_MAX)
-    {
-        config_str = elk_keycode_config_string[elk_key].config_key_string;
-    }
-    return config_str;
-}
-
-elk_key_id_t keyutils_config_string_to_elk_key_id(const char * host_config_str)
-{
-    elk_key_id_t elk_key = ELK_KEY_NONE;
-    bool found = false;
-    while(!found && ++elk_key < ELK_KEY_MAX)
-    {
-        if(strcmp(host_config_str, elk_keycode_config_string[elk_key].config_key_string) == 0)
-        {
-            found = true;
-        }
-    }
-
-    if(elk_key == ELK_KEY_MAX)
-    {
-        elk_key = ELK_KEY_NONE;
-    }
-
-    return(elk_key);
-}
