+++ src/host_abstraction_layer/keyutils.h
@@ -35,11 +35,5 @@
 *******************************************************************************/
 
 host_key_t   keyutils_get_hostkey_from_legacy_keyid(uint8_t old_key_id);
-const char * keyutils_get_hostkey_longname(host_key_t host_key);
-const char * keyutils_get_hostkey_config_string(host_key_t host_key);
-const char * keyutils_get_elkkey_longname(elk_key_id_t elk_key);
-const char * keyutils_get_elkkey_config_string(elk_key_id_t elk_key);
-
-elk_key_id_t keyutils_config_string_to_elk_key_id(const char * host_config_str);
-
+void keyhandler_refresh_elkkeys();  // Allegro 4 only.
 #endif // _KEYUTILS_H
